innovatively/requirements.txt
BeautifulSoup==3.2.1
concord-py==0.3.6
requests==2.10.0
six==1.9.0
thrift==0.9.2
zope.interface==4.1.2
